+++ frisbee_test.go
@@ -1,5 +1,5 @@
 /*
-	Copyright 2021 Loophole Labs
+	Copyright 2022 Loophole Labs
 
 	Licensed under the Apache License, Version 2.0 (the "License");
 	you may not use this file except in compliance with the License.
@@ -19,7 +19,7 @@ package frisbee_test
 import (
 	"context"
 	"github.com/loopholelabs/frisbee"
-	"github.com/loopholelabs/packet"
+	"github.com/loopholelabs/frisbee/pkg/packet"
 	"github.com/rs/zerolog"
 	"os"
 )