+++ CHANGELOG.md
@@ -7,6 +7,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [v0.3.0] - 2022-03-12 (Beta)
+
+### Changes
+
+- Moving `github.com/loopholelabs/packet` into Frisbee under `github.com/loopholelabs/frisbee/pkg/packet`
+- Added `pkg/metadata` and `pkg/content` packages to work with the `pkg/packet` package
+- Added initial version of `protoc-gen-frisbee` CLI for generating an RPC Framework with Frisbee
+
 ## [v0.2.4] - 2022-03-11 (Beta)
 
 ### Changes
@@ -157,7 +165,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 Initial Release of Frisbee
 
-[unreleased]: https://github.com/loopholelabs/frisbee/compare/v0.2.4...HEAD
+[unreleased]: https://github.com/loopholelabs/frisbee/compare/v0.3.0...HEAD
+[v0.3.0]: https://github.com/loopholelabs/frisbee/compare/v0.2.4...v0.3.0
 [v0.2.4]: https://github.com/loopholelabs/frisbee/compare/v0.2.3...v0.2.4
 [v0.2.3]: https://github.com/loopholelabs/frisbee/compare/v0.2.2...v0.2.3
 [v0.2.2]: https://github.com/loopholelabs/frisbee/compare/v0.2.1...v0.2.2